improving initial folder structure
