+++ README.md
@@ -1,1 +1,14 @@
 # Chew-Chew-Junction
+# Header
+#    Logo
+#    Navigation
+#        Home
+#        About
+#        Contact Us
+#        Cart
+# Body
+#    Search
+#    Cards Container
+# Footer
+#    Logo
+#    Links
