+++ src/components/Header.js
@@ -1,0 +1,21 @@
+import { IMG_HEAD } from "../utils/constants";
+
+const Header = () => {
+  return (
+    <div className="header">
+      <div className="logo">
+        <img src={IMG_HEAD} />
+      </div>
+      <div className="nav-items">
+        <ul>
+          <li>Home</li>
+          <li>About</li>
+          <li>Contact Us</li>
+          <li>Cart</li>
+        </ul>
+      </div>
+    </div>
+  );
+};
+
+export default Header;
