+++ src/utils/constants.js
@@ -1,0 +1,5 @@
+export const IMG_HEAD =
+  "https://i.pinimg.com/originals/01/48/66/0148662247327a7d268ce7d02b8662cf.png";
+
+export const IMG_RES =
+  "https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_1024/";
