+++ src/utils/objList.js
@@ -1,0 +1,2145 @@
+export const objList = 
+[
+        {
+          "info": {
+            "id": "59284",
+            "name": "Domino's Pizza",
+            "cloudinaryImageId": "jd3b24bmmmwsdpezahj5",
+            "locality": "Sapru Marg",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹400 for two",
+            "cuisines": [
+              "Pizzas",
+              "Italian",
+              "Pastas",
+              "Desserts"
+            ],
+            "avgRating": 4.3,
+            "feeDetails": {
+              "restaurantId": "59284",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "2456",
+            "avgRatingString": "4.3",
+            "totalRatingsString": "5K+",
+            "sla": {
+              "deliveryTime": 25,
+              "serviceability": "SERVICEABLE",
+              "slaString": "25 mins",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 02:59:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "₹125 OFF",
+              "subHeader": "ABOVE ₹999",
+              "discountTag": "FLAT DEAL"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/dominos-pizza-sapru-marg-hazratganj-lucknow-59284",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "231204",
+            "name": "Burger King",
+            "cloudinaryImageId": "e33e1d3ba7d6b2bb0d45e1001b731fcf",
+            "locality": "Mahatma Gandhi Marg",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹350 for two",
+            "cuisines": [
+              "Burgers",
+              "American"
+            ],
+            "avgRating": 4.2,
+            "feeDetails": {
+              "restaurantId": "231204",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "166",
+            "avgRatingString": "4.2",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 24,
+              "lastMileTravel": 1,
+              "serviceability": "SERVICEABLE",
+              "slaString": "24 mins",
+              "lastMileTravelString": "1.0 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 03:30:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "60% OFF",
+              "subHeader": "UPTO ₹120"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/burger-king-mahatma-gandhi-marg-hazratganj-lucknow-231204",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "405025",
+            "name": "KFC",
+            "cloudinaryImageId": "56c9ab92bd79745fd152a30fa2525426",
+            "locality": "Sahara Ganj Mall",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹400 for two",
+            "cuisines": [
+              "Burgers",
+              "Biryani",
+              "American",
+              "Snacks",
+              "Fast Food"
+            ],
+            "avgRating": 4.1,
+            "feeDetails": {
+              "restaurantId": "405025",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "547",
+            "avgRatingString": "4.1",
+            "totalRatingsString": "1K+",
+            "sla": {
+              "deliveryTime": 18,
+              "lastMileTravel": 1.3,
+              "serviceability": "SERVICEABLE",
+              "slaString": "18 mins",
+              "lastMileTravelString": "1.3 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 23:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "40% OFF",
+              "subHeader": "UPTO ₹80"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/kfc-sahara-ganj-mall-hazratganj-lucknow-405025",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "330433",
+            "name": "McDonald's",
+            "cloudinaryImageId": "08eee882fda81b68e5c36e3a5d89d8c0",
+            "locality": "Sahara Ganj Mall",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹400 for two",
+            "cuisines": [
+              "American"
+            ],
+            "avgRating": 4.1,
+            "feeDetails": {
+              "restaurantId": "330433",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "630",
+            "avgRatingString": "4.1",
+            "totalRatingsString": "5K+",
+            "sla": {
+              "deliveryTime": 28,
+              "lastMileTravel": 1.3,
+              "serviceability": "SERVICEABLE",
+              "slaString": "28 mins",
+              "lastMileTravelString": "1.3 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 22:45:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "aggregatedDiscountInfoV2": {
+              
+            },
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/mcdonalds-sahara-ganj-mall-hazratganj-lucknow-330433",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "89361",
+            "name": "La Pino'z Pizza",
+            "cloudinaryImageId": "dek7ifvgfbnddngpqn4r",
+            "locality": "Sapru Marg",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹250 for two",
+            "cuisines": [
+              "Pizzas",
+              "Pastas",
+              "Italian",
+              "Desserts",
+              "Beverages"
+            ],
+            "avgRating": 4,
+            "feeDetails": {
+              "restaurantId": "89361",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "4961",
+            "avgRatingString": "4.0",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 25,
+              "lastMileTravel": 1.7,
+              "serviceability": "SERVICEABLE",
+              "slaString": "25 mins",
+              "lastMileTravelString": "1.7 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 04:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "50% OFF",
+              "subHeader": "UPTO ₹100"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/la-pinoz-pizza-sapru-marg-hazratganj-lucknow-89361",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "84849",
+            "name": "Neelkanth Sweets",
+            "cloudinaryImageId": "s18pogvibjbumrnnabxr",
+            "locality": "Vivek Khand 3",
+            "areaName": "Gomti Nagar",
+            "costForTwo": "₹300 for two",
+            "cuisines": [
+              "Sweets",
+              "Snacks",
+              "North Indian",
+              "Chinese",
+              "South Indian",
+              "Chaat"
+            ],
+            "avgRating": 4.3,
+            "veg": true,
+            "feeDetails": {
+              "restaurantId": "84849",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 5000
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 5000
+            },
+            "parentId": "7166",
+            "avgRatingString": "4.3",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 32,
+              "lastMileTravel": 7.4,
+              "serviceability": "SERVICEABLE",
+              "slaString": "32 mins",
+              "lastMileTravelString": "7.4 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 22:01:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "₹125 OFF",
+              "subHeader": "ABOVE ₹199",
+              "discountTag": "FLAT DEAL"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/neelkanth-sweets-vivek-khand-3-gomti-nagar-lucknow-84849",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "60240",
+            "name": "Pizza Hut",
+            "cloudinaryImageId": "2b4f62d606d1b2bfba9ba9e5386fabb7",
+            "locality": "MG Marg",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹350 for two",
+            "cuisines": [
+              "Pizzas"
+            ],
+            "avgRating": 3.8,
+            "feeDetails": {
+              "restaurantId": "60240",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "721",
+            "avgRatingString": "3.8",
+            "totalRatingsString": "5K+",
+            "sla": {
+              "deliveryTime": 23,
+              "lastMileTravel": 0.9,
+              "serviceability": "SERVICEABLE",
+              "slaString": "23 mins",
+              "lastMileTravelString": "0.9 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 04:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "50% OFF",
+              "subHeader": "UPTO ₹100"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/pizza-hut-mg-marg-hazratganj-lucknow-60240",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "64649",
+            "name": "Tunday Kababi (Aminabad)",
+            "cloudinaryImageId": "uws6wm1xndcgtox6moif",
+            "locality": "Khayali Ganj",
+            "areaName": "Aminabad",
+            "costForTwo": "₹200 for two",
+            "cuisines": [
+              "Mughlai"
+            ],
+            "avgRating": 4.3,
+            "feeDetails": {
+              "restaurantId": "64649",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 2300
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 2300
+            },
+            "parentId": "218444",
+            "avgRatingString": "4.3",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 30,
+              "lastMileTravel": 3,
+              "serviceability": "SERVICEABLE",
+              "slaString": "30 mins",
+              "lastMileTravelString": "3.0 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 23:30:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "50% OFF",
+              "subHeader": "UPTO ₹100"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/tunday-kababi-aminabad-khayali-ganj-aminabad-lucknow-64649",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "229723",
+            "name": "Faasos - Wraps & Rolls",
+            "cloudinaryImageId": "af33b81798b11deba338e94b7585d348",
+            "locality": "Habibulla Estate Road",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹200 for two",
+            "cuisines": [
+              "Kebabs",
+              "Fast Food",
+              "Snacks",
+              "North Indian",
+              "American",
+              "Healthy Food",
+              "Desserts",
+              "Beverages"
+            ],
+            "avgRating": 4.2,
+            "feeDetails": {
+              "restaurantId": "229723",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "21809",
+            "avgRatingString": "4.2",
+            "totalRatingsString": "5K+",
+            "sla": {
+              "deliveryTime": 25,
+              "lastMileTravel": 1.4,
+              "serviceability": "SERVICEABLE",
+              "slaString": "25 mins",
+              "lastMileTravelString": "1.4 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 03:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "₹125 OFF",
+              "subHeader": "ABOVE ₹249",
+              "discountTag": "FLAT DEAL"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/faasos-wraps-and-rolls-habibulla-estate-road-hazratganj-lucknow-229723",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "59654",
+            "name": "Subway",
+            "cloudinaryImageId": "1ace5fa65eff3e1223feb696c956b38b",
+            "locality": "Sapru Marg",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹350 for two",
+            "cuisines": [
+              "Healthy Food",
+              "Salads",
+              "Snacks",
+              "Desserts",
+              "Beverages"
+            ],
+            "avgRating": 4.1,
+            "feeDetails": {
+              "restaurantId": "59654",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "2",
+            "avgRatingString": "4.1",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 20,
+              "lastMileTravel": 0.9,
+              "serviceability": "SERVICEABLE",
+              "slaString": "20 mins",
+              "lastMileTravelString": "0.9 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 04:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "50% OFF",
+              "subHeader": "UPTO ₹100"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/subway-sapru-marg-hazratganj-lucknow-59654",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "101055",
+            "name": "Wow! Momo",
+            "cloudinaryImageId": "0984acc0ed7b914206dbbcb90297becc",
+            "locality": "Sapru Marg",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹300 for two",
+            "cuisines": [
+              "Tibetan",
+              "Healthy Food",
+              "Asian",
+              "Chinese",
+              "Snacks",
+              "Continental",
+              "Desserts",
+              "Beverages"
+            ],
+            "avgRating": 3.8,
+            "feeDetails": {
+              "restaurantId": "101055",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "1776",
+            "avgRatingString": "3.8",
+            "totalRatingsString": "5K+",
+            "sla": {
+              "deliveryTime": 25,
+              "lastMileTravel": 0.9,
+              "serviceability": "SERVICEABLE",
+              "slaString": "25 mins",
+              "lastMileTravelString": "0.9 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 03:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "₹125 OFF",
+              "subHeader": "ABOVE ₹249",
+              "discountTag": "FLAT DEAL"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/wow-momo-sapru-marg-hazratganj-lucknow-101055",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "61689",
+            "name": "Bikanervala",
+            "cloudinaryImageId": "m4thufeyptbyg2qxcg8m",
+            "locality": "Sector 5",
+            "areaName": "Indira Nagar",
+            "costForTwo": "₹200 for two",
+            "cuisines": [
+              "South Indian",
+              "North Indian"
+            ],
+            "avgRating": 4.3,
+            "veg": true,
+            "feeDetails": {
+              "restaurantId": "61689",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 4000
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 4000
+            },
+            "parentId": "45936",
+            "avgRatingString": "4.3",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 34,
+              "lastMileTravel": 5,
+              "serviceability": "SERVICEABLE",
+              "slaString": "34 mins",
+              "lastMileTravelString": "5.0 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 23:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "20% OFF",
+              "subHeader": "UPTO ₹50"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/bikanervala-sector-5-indira-nagar-lucknow-61689",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "229725",
+            "name": "Oven Story Pizza",
+            "cloudinaryImageId": "ab979bffbd658e74de650a15ca0092a3",
+            "locality": "Habibulla Estate Road",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹400 for two",
+            "cuisines": [
+              "Pizzas",
+              "Pastas",
+              "Italian",
+              "Desserts",
+              "Beverages"
+            ],
+            "avgRating": 4,
+            "feeDetails": {
+              "restaurantId": "229725",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "3534",
+            "avgRatingString": "4.0",
+            "totalRatingsString": "1K+",
+            "sla": {
+              "deliveryTime": 28,
+              "lastMileTravel": 1.4,
+              "serviceability": "SERVICEABLE",
+              "slaString": "28 mins",
+              "lastMileTravelString": "1.4 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 03:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "₹125 OFF",
+              "subHeader": "ABOVE ₹1099",
+              "discountTag": "FLAT DEAL"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/oven-story-pizza-habibulla-estate-road-hazratganj-lucknow-229725",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "65085",
+            "name": "Mughal Zaika Aminabad",
+            "cloudinaryImageId": "yood4aumr1wbxed1odxk",
+            "locality": "Naaz Cinema Road",
+            "areaName": "Aminabad",
+            "costForTwo": "₹300 for two",
+            "cuisines": [
+              "Mughlai"
+            ],
+            "avgRating": 4.1,
+            "feeDetails": {
+              "restaurantId": "65085",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 2300
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 2300
+            },
+            "parentId": "140984",
+            "avgRatingString": "4.1",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 32,
+              "lastMileTravel": 3,
+              "serviceability": "SERVICEABLE",
+              "slaString": "32 mins",
+              "lastMileTravelString": "3.0 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 23:59:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "₹125 OFF",
+              "subHeader": "ABOVE ₹199",
+              "discountTag": "FLAT DEAL"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/mughal-zaika-aminabad-naaz-cinema-road-aminabad-lucknow-65085",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "82387",
+            "name": "Red Dragon (Chopstix)",
+            "cloudinaryImageId": "f6etwqizmuaxxe21vxmu",
+            "locality": "Sector C",
+            "areaName": "Mahanagar",
+            "costForTwo": "₹250 for two",
+            "cuisines": [
+              "Chinese",
+              "Arabian"
+            ],
+            "avgRating": 3.8,
+            "feeDetails": {
+              "restaurantId": "82387",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 2800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 2800
+            },
+            "parentId": "168369",
+            "avgRatingString": "3.8",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 29,
+              "lastMileTravel": 4.3,
+              "serviceability": "SERVICEABLE",
+              "slaString": "29 mins",
+              "lastMileTravelString": "4.3 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 02:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "50% OFF",
+              "subHeader": "UPTO ₹100"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/red-dragon-chopstix-sector-c-mahanagar-lucknow-82387",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "63930",
+            "name": "The Bon Bon Bakers",
+            "cloudinaryImageId": "2c761e85049f81355f6bf864c97666dd",
+            "locality": "Sector A",
+            "areaName": "Aliganj",
+            "costForTwo": "₹300 for two",
+            "cuisines": [
+              "Desserts",
+              "Pizzas",
+              "Snacks",
+              "Italian",
+              "Beverages"
+            ],
+            "avgRating": 4.3,
+            "feeDetails": {
+              "restaurantId": "63930",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 4500
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 4500
+            },
+            "parentId": "205683",
+            "avgRatingString": "4.3",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 34,
+              "lastMileTravel": 6.9,
+              "serviceability": "SERVICEABLE",
+              "slaString": "34 mins",
+              "lastMileTravelString": "6.9 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 22:30:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "60% OFF",
+              "subHeader": "UPTO ₹120"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/the-bon-bon-bakers-sector-a-aliganj-lucknow-63930",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "165050",
+            "name": "Mahalaxmi Sweets & Restaurant",
+            "cloudinaryImageId": "oqk0mierg0rlhuxoa3ok",
+            "locality": "Purania Road",
+            "areaName": "Aliganj",
+            "costForTwo": "₹300 for two",
+            "cuisines": [
+              "Indian",
+              "North Indian"
+            ],
+            "avgRating": 4.2,
+            "veg": true,
+            "feeDetails": {
+              "restaurantId": "165050",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 4500
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 4500
+            },
+            "parentId": "262175",
+            "avgRatingString": "4.2",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 34,
+              "lastMileTravel": 6.9,
+              "serviceability": "SERVICEABLE",
+              "slaString": "34 mins",
+              "lastMileTravelString": "6.9 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 22:30:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "60% OFF",
+              "subHeader": "UPTO ₹120"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/mahalaxmi-sweets-and-restaurant-purania-road-aliganj-lucknow-165050",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "59530",
+            "name": "The Cherry Tree Bakery & Cafe",
+            "cloudinaryImageId": "8b396b7aca4b55e53c647e68c2b95a93",
+            "locality": "Habibullah Estate Road",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹250 for two",
+            "cuisines": [
+              "Bakery",
+              "Pastas",
+              "Pizzas",
+              "Continental",
+              "Snacks"
+            ],
+            "avgRating": 4.3,
+            "feeDetails": {
+              "restaurantId": "59530",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "6261",
+            "avgRatingString": "4.3",
+            "totalRatingsString": "5K+",
+            "sla": {
+              "deliveryTime": 22,
+              "lastMileTravel": 1.2,
+              "serviceability": "SERVICEABLE",
+              "slaString": "22 mins",
+              "lastMileTravelString": "1.2 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 23:45:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "40% OFF",
+              "subHeader": "UPTO ₹80"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/the-cherry-tree-bakery-and-cafe-habibullah-estate-road-hazratganj-lucknow-59530",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "65082",
+            "name": "Classic Radhey Sweets",
+            "cloudinaryImageId": "hogs5exnbgrll77362qy",
+            "locality": "Mandir Marg",
+            "areaName": "Mahanagar",
+            "costForTwo": "₹300 for two",
+            "cuisines": [
+              "Sweets",
+              "Navratri Special"
+            ],
+            "avgRating": 4.2,
+            "veg": true,
+            "feeDetails": {
+              "restaurantId": "65082",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 2300
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 2300
+            },
+            "parentId": "63739",
+            "avgRatingString": "4.2",
+            "totalRatingsString": "10K+",
+            "sla": {
+              "deliveryTime": 26,
+              "lastMileTravel": 3.9,
+              "serviceability": "SERVICEABLE",
+              "slaString": "26 mins",
+              "lastMileTravelString": "3.9 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-23 23:30:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "50% OFF",
+              "subHeader": "UPTO ₹100"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/classic-radhey-sweets-mandir-marg-mahanagar-lucknow-65082",
+            "type": "WEBLINK"
+          }
+        },
+        {
+          "info": {
+            "id": "63376",
+            "name": "Keventers - Milkshakes & Desserts",
+            "cloudinaryImageId": "af8216b15abde417bb0e1d8fc44fb877",
+            "locality": "Sapru Marg",
+            "areaName": "Hazratganj",
+            "costForTwo": "₹200 for two",
+            "cuisines": [
+              "Beverages",
+              "Ice Cream",
+              "Desserts",
+              "Healthy Food"
+            ],
+            "avgRating": 4.3,
+            "veg": true,
+            "feeDetails": {
+              "restaurantId": "63376",
+              "fees": [
+                {
+                  "name": "BASE_DISTANCE",
+                  "fee": 1800
+                },
+                {
+                  "name": "BASE_TIME"
+                },
+                {
+                  "name": "ANCILLARY_SURGE_FEE"
+                }
+              ],
+              "totalFee": 1800
+            },
+            "parentId": "268997",
+            "avgRatingString": "4.3",
+            "totalRatingsString": "5K+",
+            "sla": {
+              "deliveryTime": 18,
+              "lastMileTravel": 0.9,
+              "serviceability": "SERVICEABLE",
+              "slaString": "18 mins",
+              "lastMileTravelString": "0.9 km",
+              "iconType": "ICON_TYPE_EMPTY"
+            },
+            "availability": {
+              "nextCloseTime": "2023-07-24 04:00:00",
+              "opened": true
+            },
+            "badges": {
+              
+            },
+            "isOpen": true,
+            "type": "F",
+            "badgesV2": {
+              "entityBadges": {
+                "imageBased": {
+                  
+                },
+                "textBased": {
+                  
+                },
+                "textExtendedBadges": {
+                  
+                }
+              }
+            },
+            "aggregatedDiscountInfoV3": {
+              "header": "40% OFF",
+              "subHeader": "UPTO ₹80"
+            },
+            "orderabilityCommunication": {
+              "title": {
+                
+              },
+              "subTitle": {
+                
+              },
+              "message": {
+                
+              },
+              "customIcon": {
+                
+              }
+            },
+            "differentiatedUi": {
+              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+              "differentiatedUiMediaDetails": {
+                "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+                "lottie": {
+                  
+                },
+                "video": {
+                  
+                }
+              }
+            },
+            "reviewsSummary": {
+              
+            },
+            "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT"
+          },
+          "analytics": {
+            
+          },
+          "cta": {
+            "link": "https://www.swiggy.com/restaurants/keventers-milkshakes-and-desserts-sapru-marg-hazratganj-lucknow-63376",
+            "type": "WEBLINK"
+          }
+        }
+];
